+++ lib/leitura_entrada.h
@@ -1,1 +1,1 @@
-void leia_entrada(char*, int*, int*, int*, int*, int*, int*, int*, unsigned int*, double*, double*);
+void leia_entrada(char*, int*, int*, int*, int*, int*, int*, int*, unsigned int*, float*, float*);
